Adds docs and tests for hash function feature
