+++ pkg/js/parse_tests/051-HASH.js
@@ -1,0 +1,1 @@
+D(HASH("SHA1", "abc"), "reg")
