+++ pkg/js/parse_tests/051-HASH.json
@@ -1,0 +1,12 @@
+{
+  "registrars": [],
+  "dns_providers": [],
+  "domains": [
+    {
+      "name": "a9993e364706816aba3e25717850c26c9cd0d89d",
+      "registrar": "reg",
+      "dnsProviders": {},
+      "records": []
+    }
+  ]
+}
